# backend/internal/transport/http/openapi/spec.yaml
openapi: 3.0.3
info:
  title: StreamPulse API
  version: 1.0.0
  description: |
    Real-time audio streaming platform — RNCP 38822 Bloc 3.
    The streaming endpoints (`/streams/:id/publish`, `/streams/:id/listen`)
    use chunked HTTP transfer rather than WebSocket / WebRTC.
  contact:
    email: contact@ecole-decode.fr
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://staging.streampulse.example
    description: Staging
  - url: https://streampulse.example
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        username: { type: string }
        role:
          type: string
          enum: [anonymous, user, broadcaster, admin]
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: '#/components/schemas/UserResponse' }
    RegisterRequest:
      type: object
      required: [email, username, password]
      properties:
        email: { type: string, format: email }
        username: { type: string, minLength: 3, maxLength: 32 }
        password: { type: string, minLength: 8, maxLength: 128, format: password }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, format: password }
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegisterRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: Invalid input
        '409':
          description: Email or username taken
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Sign in
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh JWT (sliding window)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
...